refactor(src): refactor the file structure for better organization
the entire src structure has been refactored by splitting items into individual files

BREAKING CHANGE: this restructure will potentially be a breaking change for projects that use deep
imports. if your project uses deep imports please refactor them to the new structure.
