+++ .editorconfig
@@ -1,0 +1,9 @@
+root = true
+
+[{src,scripts}/**.{ts,json,js}]
+end_of_line = lf
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
+indent_style = space
+indent_size = 2
