+++ src/constants/Capability.ts
@@ -1,0 +1,20 @@
+import { NativeModules } from 'react-native';
+const { TrackPlayerModule: TrackPlayer } = NativeModules;
+
+export enum Capability {
+  Play = TrackPlayer.CAPABILITY_PLAY,
+  PlayFromId = TrackPlayer.CAPABILITY_PLAY_FROM_ID,
+  PlayFromSearch = TrackPlayer.CAPABILITY_PLAY_FROM_SEARCH,
+  Pause = TrackPlayer.CAPABILITY_PAUSE,
+  Stop = TrackPlayer.CAPABILITY_STOP,
+  SeekTo = TrackPlayer.CAPABILITY_SEEK_TO,
+  Skip = TrackPlayer.CAPABILITY_SKIP,
+  SkipToNext = TrackPlayer.CAPABILITY_SKIP_TO_NEXT,
+  SkipToPrevious = TrackPlayer.CAPABILITY_SKIP_TO_PREVIOUS,
+  JumpForward = TrackPlayer.CAPABILITY_JUMP_FORWARD,
+  JumpBackward = TrackPlayer.CAPABILITY_JUMP_BACKWARD,
+  SetRating = TrackPlayer.CAPABILITY_SET_RATING,
+  Like = TrackPlayer.CAPABILITY_LIKE,
+  Dislike = TrackPlayer.CAPABILITY_DISLIKE,
+  Bookmark = TrackPlayer.CAPABILITY_BOOKMARK,
+}
