+++ src/constants/TrackType.ts
@@ -1,0 +1,6 @@
+export enum TrackType {
+  Default = 'default',
+  Dash = 'dash',
+  HLS = 'hls',
+  SmoothStreaming = 'smoothstreaming',
+}
