+++ src/interfaces/FeedbackOptions.ts
@@ -1,0 +1,7 @@
+export interface FeedbackOptions {
+  /** Marks wether the option should be marked as active or "done" */
+  isActive: boolean;
+
+  /** The title to give the action (relevant for iOS) */
+  title: string;
+}
