+++ src/interfaces/MetadataOptions.ts
@@ -1,0 +1,4 @@
+import type { UpdateOptions } from './UpdateOptions';
+
+/** @deprecated use UpdateOptions instead */
+export type MetadataOptions = UpdateOptions;
