+++ src/interfaces/NowPlayingMetadata.ts
@@ -1,0 +1,5 @@
+import type { TrackMetadataBase } from './TrackMetadataBase';
+
+export interface NowPlayingMetadata extends TrackMetadataBase {
+  elapsedTime?: number;
+}
