+++ src/interfaces/PlaybackState.ts
@@ -1,0 +1,11 @@
+import { State } from '../constants';
+import type { PlaybackErrorEvent } from './events';
+
+export type PlaybackState =
+  | {
+      state: Exclude<State, State.Error>;
+    }
+  | {
+      state: State.Error;
+      error: PlaybackErrorEvent;
+    };
