+++ src/interfaces/PlayerOptions.ts
@@ -1,0 +1,81 @@
+import type {
+  IOSCategory,
+  IOSCategoryMode,
+  IOSCategoryOptions,
+  AndroidAudioContentType,
+} from '../constants';
+
+export interface PlayerOptions {
+  /**
+   * Minimum time in seconds that needs to be buffered.
+   */
+  minBuffer?: number;
+  /**
+   * Maximum time in seconds that needs to be buffered
+   */
+  maxBuffer?: number;
+  /**
+   * Time in seconds that should be kept in the buffer behind the current playhead time.
+   */
+  backBuffer?: number;
+  /**
+   * Minimum time in seconds that needs to be buffered to start playing.
+   */
+  playBuffer?: number;
+  /**
+   * Maximum cache size in kilobytes.
+   */
+  maxCacheSize?: number;
+  /**
+   * [AVAudioSession.Category](https://developer.apple.com/documentation/avfoundation/avaudiosession/1616615-category)
+   * for iOS. Sets on `play()`.
+   */
+  iosCategory?: IOSCategory;
+  /**
+   * (iOS only) The audio session mode, together with the audio session category,
+   * indicates to the system how you intend to use audio in your app. You can use
+   * a mode to configure the audio system for specific use cases such as video
+   * recording, voice or video chat, or audio analysis.
+   * Sets on `play()`.
+   *
+   * See https://developer.apple.com/documentation/avfoundation/avaudiosession/1616508-mode
+   */
+  iosCategoryMode?: IOSCategoryMode;
+  /**
+   * [AVAudioSession.CategoryOptions](https://developer.apple.com/documentation/avfoundation/avaudiosession/1616503-categoryoptions) for iOS.
+   * Sets on `play()`.
+   */
+  iosCategoryOptions?: IOSCategoryOptions[];
+  /**
+   * (Android only) The audio content type indicates to the android system how
+   * you intend to use audio in your app.
+   *
+   * With `autoHandleInterruptions: true` and
+   * `androidContentType: AndroidAudioContentType.Speech`, the audio will be
+   * paused during short interruptions, such as when a message arrives.
+   * Otherwise the playback volume is reduced while the notification is playing.
+   *
+   * @default AndroidAudioContentType.Music
+   */
+  androidAudioContentType?: AndroidAudioContentType;
+  /**
+   * Indicates whether the player should automatically delay playback in order to minimize stalling.
+   * Defaults to `true`.
+   * @deprecated This option has been nominated for removal in a future version
+   * of RNTP. If you have this set to `true`, you can safely remove this from
+   * the options. If you are setting this to `false` and have a reason for that,
+   * please post a comment in the following discussion: https://github.com/doublesymmetry/react-native-track-player/pull/1695
+   * and describe why you are doing so.
+   */
+  waitForBuffer?: boolean;
+  /**
+   * Indicates whether the player should automatically update now playing metadata data in control center / notification.
+   * Defaults to `true`.
+   */
+  autoUpdateMetadata?: boolean;
+  /**
+   * Indicates whether the player should automatically handle audio interruptions.
+   * Defaults to `false`.
+   */
+  autoHandleInterruptions?: boolean;
+}
