+++ src/interfaces/SleepTimer.ts
@@ -1,0 +1,4 @@
+export type SleepTimer =
+  | { time: number }
+  | { sleepWhenPlayedToEnd: boolean }
+  | null;
