+++ src/interfaces/TrackMetadataBase.ts
@@ -1,0 +1,28 @@
+import type { ResourceObject } from './ResourceObject';
+import type { RatingType } from '../constants';
+
+export interface TrackMetadataBase {
+  /** The track title */
+  title?: string;
+  /** The track album */
+  album?: string;
+  /** The track artist */
+  artist?: string;
+  /** The track duration in seconds */
+  duration?: number;
+  /** The track artwork */
+  artwork?: string | ResourceObject;
+  /** track description */
+  description?: string;
+  /** The track genre */
+  genre?: string;
+  /** The track release date in [RFC 3339](https://www.ietf.org/rfc/rfc3339.txt) */
+  date?: string;
+  /** The track rating */
+  rating?: RatingType;
+  /**
+   * (iOS only) Whether the track is presented in the control center as being
+   * live
+   **/
+  isLiveStream?: boolean;
+}
