+++ src/interfaces/events/EventPayloadByEvent.ts
@@ -1,0 +1,48 @@
+import { Event } from '../../constants';
+
+import type { PlaybackState } from '../PlaybackState';
+import type { PlaybackErrorEvent } from './PlaybackErrorEvent';
+import type { PlaybackQueueEndedEvent } from './PlaybackQueueEndedEvent';
+import type { PlaybackTrackChangedEvent } from './PlaybackTrackChangedEvent';
+import type { PlaybackActiveTrackChangedEvent } from './PlaybackActiveTrackChangedEvent';
+import type { PlaybackMetadataReceivedEvent } from './PlaybackMetadataReceivedEvent';
+import type { PlaybackPlayWhenReadyChangedEvent } from './PlaybackPlayWhenReadyChangedEvent';
+import type { PlaybackProgressUpdatedEvent } from './PlaybackProgressUpdatedEvent';
+import type { SleepTimerChangedEvent } from './SleepTimerChangedEvent';
+import type { RemotePlayIdEvent } from './RemotePlayIdEvent';
+import type { RemotePlaySearchEvent } from './RemotePlaySearchEvent';
+import type { RemoteSkipEvent } from './RemoteSkipEvent';
+import type { RemoteJumpForwardEvent } from './RemoteJumpForwardEvent';
+import type { RemoteJumpBackwardEvent } from './RemoteJumpBackwardEvent';
+import type { RemoteSeekEvent } from './RemoteSeekEvent';
+import type { RemoteSetRatingEvent } from './RemoteSetRatingEvent';
+import type { RemoteDuckEvent } from './RemoteDuckEvent';
+
+export interface EventPayloadByEvent {
+  [Event.PlaybackState]: PlaybackState;
+  [Event.PlaybackError]: PlaybackErrorEvent;
+  [Event.PlaybackQueueEnded]: PlaybackQueueEndedEvent;
+  [Event.PlaybackTrackChanged]: PlaybackTrackChangedEvent;
+  [Event.PlaybackActiveTrackChanged]: PlaybackActiveTrackChangedEvent;
+  [Event.PlaybackMetadataReceived]: PlaybackMetadataReceivedEvent;
+  [Event.PlaybackPlayWhenReadyChanged]: PlaybackPlayWhenReadyChangedEvent;
+  [Event.PlaybackProgressUpdated]: PlaybackProgressUpdatedEvent;
+  [Event.SleepTimerChanged]: SleepTimerChangedEvent;
+  [Event.SleepTimerComplete]: never;
+  [Event.RemotePlay]: never;
+  [Event.RemotePlayId]: RemotePlayIdEvent;
+  [Event.RemotePlaySearch]: RemotePlaySearchEvent;
+  [Event.RemotePause]: never;
+  [Event.RemoteStop]: never;
+  [Event.RemoteSkip]: RemoteSkipEvent;
+  [Event.RemoteNext]: never;
+  [Event.RemotePrevious]: never;
+  [Event.RemoteJumpForward]: RemoteJumpForwardEvent;
+  [Event.RemoteJumpBackward]: RemoteJumpBackwardEvent;
+  [Event.RemoteSeek]: RemoteSeekEvent;
+  [Event.RemoteSetRating]: RemoteSetRatingEvent;
+  [Event.RemoteDuck]: RemoteDuckEvent;
+  [Event.RemoteLike]: never;
+  [Event.RemoteDislike]: never;
+  [Event.RemoteBookmark]: never;
+}
