+++ src/interfaces/events/PlaybackErrorEvent.ts
@@ -1,0 +1,6 @@
+export interface PlaybackErrorEvent {
+  /** The error code */
+  code: string;
+  /** The error message */
+  message: string;
+}
