+++ src/interfaces/events/PlaybackPlayWhenReadyChangedEvent.ts
@@ -1,0 +1,4 @@
+export interface PlaybackPlayWhenReadyChangedEvent {
+  /** Whether the player will play when it is ready to do so. */
+  playWhenReady: boolean;
+}
