+++ src/interfaces/events/RemoteJumpBackwardEvent.ts
@@ -1,0 +1,7 @@
+export interface RemoteJumpBackwardEvent {
+  /**
+   * The number of seconds to jump backward.
+   * See https://react-native-track-player.js.org/docs/api/events#remotejumpbackward
+   **/
+  interval: number;
+}
