+++ src/interfaces/events/RemoteJumpForwardEvent.ts
@@ -1,0 +1,7 @@
+export interface RemoteJumpForwardEvent {
+  /**
+   * The number of seconds to jump forward.
+   * See https://react-native-track-player.js.org/docs/api/events#remotejumpforward
+   **/
+  interval: number;
+}
