+++ src/interfaces/events/RemotePlayIdEvent.ts
@@ -1,0 +1,4 @@
+export interface RemotePlayIdEvent {
+  /** The track id */
+  id: string;
+}
