+++ src/interfaces/events/RemoteSeekEvent.ts
@@ -1,0 +1,4 @@
+export interface RemoteSeekEvent {
+  /** The position to seek to in seconds. */
+  position: number;
+}
