+++ src/interfaces/events/RemoteSetRatingEvent.ts
@@ -1,0 +1,5 @@
+import type { RatingType } from '../../constants';
+
+export interface RemoteSetRatingEvent {
+  rating: RatingType;
+}
