+++ src/interfaces/events/SleepTimerChangedEvent.ts
@@ -1,0 +1,3 @@
+import type { SleepTimer } from '../SleepTimer';
+
+export type SleepTimerChangedEvent = SleepTimer;
