+++ src/interfaces/index.ts
@@ -1,0 +1,14 @@
+export * from './AndroidOptions';
+export * from './FeedbackOptions';
+export * from './MetadataOptions';
+export * from './NowPlayingMetadata';
+export * from './PlaybackState';
+export * from './PlayerOptions';
+export * from './Progress';
+export * from './ResourceObject';
+export * from './ServiceHandler';
+export * from './SleepTimer';
+export * from './Track';
+export * from './TrackMetadataBase';
+export * from './UpdateOptions';
+export * from './events';
